[refactor] UserCoupon 패키지 분리 ( #12 )
