+++ src/main/java/com/coupon/issuecouponservice/domain/user/User.java
@@ -1,7 +1,7 @@
 package com.coupon.issuecouponservice.domain.user;
 
 import com.coupon.issuecouponservice.domain.common.Timestamped;
-import com.coupon.issuecouponservice.domain.coupon.UserCoupon;
+import com.coupon.issuecouponservice.domain.user_coupon.UserCoupon;
 import com.coupon.issuecouponservice.dto.request.user.UserModificationParam;
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
