+++ src/main/java/com/coupon/issuecouponservice/domain/user_coupon/UserCoupon.java
@@ -1,5 +1,6 @@
-package com.coupon.issuecouponservice.domain.coupon;
+package com.coupon.issuecouponservice.domain.user_coupon;
 
+import com.coupon.issuecouponservice.domain.coupon.Coupon;
 import com.coupon.issuecouponservice.domain.user.User;
 import io.hypersistence.utils.hibernate.id.Tsid;
 import jakarta.persistence.*;
